+++ vm/src/air_private_input.rs
@@ -369,7 +369,7 @@ mod tests {
 
     #[test]
     fn serialize_air_private_input_small_layout_only_builtins() {
-        let config = crate::cairo_run::CairoRunConfig {
+        let config = crate::cairo_run::Cairo0RunConfig {
             proof_mode: true,
             fill_holes: true,
             relocate_mem: true,